Updated data types for efficiency

+++ insert_pollution_data.sql
@@ -4,8 +4,8 @@ CREATE DEFINER=`root`@`localhost` PROCEDURE `insert_pollution_data`(
     country VARCHAR(45),
     state VARCHAR(45),
     city VARCHAR(45),
-    us_aqi INT,
-    cn_aqi INT
+    us_aqi SMALLINT,
+    cn_aqi SMALLINT
 )
 BEGIN
 	INSERT INTO pollution
